lab2/user_service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Сервис для пользователей"
authors = [
    {name = "KiriruZaytsev",email = "kirillwestcoast@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
